Navigation functionality update.
Added a bottom navigation bar and a main screen that controls the navigation with state management.
Also made changes to the main control file to check if there is a user logged in already or not. If not it starts in the login screen; if yes, it starts in the landing screen where it fetches the weather automatically.